[java] New Rule UnsynchronizedStaticFormatterRule This rule replaces UnsynchronizedStaticDateFormatterRule
