+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/multithreading/UnsynchronizedStaticDateFormatterRule.java
@@ -6,16 +6,6 @@ package net.sourceforge.pmd.lang.java.rule.multithreading;
 
 import java.text.DateFormat;
 
-import net.sourceforge.pmd.lang.ast.Node;
-import net.sourceforge.pmd.lang.java.ast.ASTClassOrInterfaceType;
-import net.sourceforge.pmd.lang.java.ast.ASTFieldDeclaration;
-import net.sourceforge.pmd.lang.java.ast.ASTMethodDeclaration;
-import net.sourceforge.pmd.lang.java.ast.ASTSynchronizedStatement;
-import net.sourceforge.pmd.lang.java.ast.ASTVariableDeclaratorId;
-import net.sourceforge.pmd.lang.java.rule.AbstractJavaRule;
-import net.sourceforge.pmd.lang.java.typeresolution.TypeHelper;
-import net.sourceforge.pmd.lang.symboltable.NameOccurrence;
-
 /**
  * Using a DateFormatter (SimpleDateFormatter) which is static can cause
  * unexpected results when used in a multi-threaded environment. This rule will
@@ -29,39 +19,12 @@ import net.sourceforge.pmd.lang.symboltable.NameOccurrence;
  * 
  * @author Allan Caplan
  * @see <a href="https://sourceforge.net/p/pmd/feature-requests/226/">feature #226 Check for SimpleDateFormat as singleton?</a>
+ * @deprecated This rule is being replaced by {@link UnsynchronizedStaticFormatterRule}. The rule will be removed with PMD 7.0.0.
  */
-public class UnsynchronizedStaticDateFormatterRule extends AbstractJavaRule {
+@Deprecated
+public class UnsynchronizedStaticDateFormatterRule extends UnsynchronizedStaticFormatterRule {
 
     public UnsynchronizedStaticDateFormatterRule() {
-        addRuleChainVisit(ASTFieldDeclaration.class);
-    }
-
-    @Override
-    public Object visit(ASTFieldDeclaration node, Object data) {
-        if (!node.isStatic()) {
-            return data;
-        }
-        ASTClassOrInterfaceType cit = node.getFirstDescendantOfType(ASTClassOrInterfaceType.class);
-        if (cit == null || !TypeHelper.isA(cit, DateFormat.class)) {
-            return data;
-        }
-
-        ASTVariableDeclaratorId var = node.getFirstDescendantOfType(ASTVariableDeclaratorId.class);
-        for (NameOccurrence occ : var.getUsages()) {
-            Node n = occ.getLocation();
-            if (n.getFirstParentOfType(ASTSynchronizedStatement.class) != null) {
-                continue;
-            }
-            // ignore usages, that don't call a method.
-            if (!n.getImage().contains(".")) {
-                continue;
-            }
-
-            ASTMethodDeclaration method = n.getFirstParentOfType(ASTMethodDeclaration.class);
-            if (method != null && !method.isSynchronized()) {
-                addViolation(data, n);
-            }
-        }
-        return data;
+        super(DateFormat.class);
     }
 }
