+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/rule/multithreading/UnsynchronizedStaticFormatterTest.java
@@ -1,0 +1,11 @@
+/**
+ * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
+ */
+
+package net.sourceforge.pmd.lang.java.rule.multithreading;
+
+import net.sourceforge.pmd.testframework.PmdRuleTst;
+
+public class UnsynchronizedStaticFormatterTest extends PmdRuleTst {
+    // no additional unit tests
+}
